Implement stuff for concurrency.
This took some refactoring too for best effect.
