+++ Cargo.toml
@@ -12,4 +12,5 @@ license = "MIT/Apache-2.0"
 
 [features]
 clone = []
+concurrent = []
 nightly = []
